refactoring + added whisper - need sudo for keypress
