+++ requirements.txt
@@ -1,3 +1,4 @@
 TTS
+openai-whisper
 keyboard
 pyaudio
